requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.6
pydantic==2.8.2
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==10.4.0
openai==1.40.2
duckdb==1.1.1
pydantic-settings==2.4.0
python-multipart==0.0.9